refactor: extract api-keys/mcp/budget sections
Move three more settings sections into per-section components under
components/settings/.
